Cambios
-No hay responsable
-Comentarios En controladores
-Imagenes arreglado

+++ app/Http/Controllers/EquipoController.php
@@ -9,19 +9,18 @@ use Illuminate\Support\Facades\File;
 class EquipoController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Muestra el inventario con tipo de inventario 'Equipo'.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
         $equipos = Inventario::where('tipoinventario', '=', 'Equipo')->get() ;
-        //dd($equipos);
         return view('equipo.index',compact('equipos'));
     }
 
     /**
-     * Show the form for creating a new resource.
+     * Muestra la creación de un nuevo equipo.
      *
      * @return \Illuminate\Http\Response
      */
@@ -29,38 +28,43 @@ class EquipoController extends Controller
     {
         $equipo = new Equipo();
         $numero = Inventario::max('idinventario') +1;
-
         return view('equipo.create', compact('equipo'), compact('numero'));
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Guarda la creación de un nuevo equipo.
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
+        //Revisa si hay una imagen insertandose
         if ($request->hasfile('fotoinventario')) {
-
             $file = $request->file('fotoinventario');
             $extention = $file->getClientOriginalExtension();
-            $filename = '/'.time().'.'.$extention;
+            $filename = time().'.'.$extention;
             $file->move(public_path().'/Imagenes', $filename);
-            //dd($filename);
+    
+            Inventario::insert([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
+                'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+            ]);  
         }
-        Inventario::insert([
-            'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
-            'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
-            'estadoinventario' => 'En Buenas Condiciones', 'informacioninventario' => $request->informacioninventario
-        
-        ]);
+        //Si no, se incluye una imagen por defecto para mostrar.
+        else {
+            Inventario::insert([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => 'sinimagen.jpg',
+                'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+            ]);
+        }
         return redirect()->route('equipo.index');
-        //dd($equipo);
     }
 
     /**
-     * Display the specified resource.
+     * Muestra el equipo en específico.
      *
      * @param  \App\Models\Equipo  $equipo
      * @return \Illuminate\Http\Response
@@ -68,25 +72,23 @@ class EquipoController extends Controller
     public function show(Inventario $equipo)
     {
         $equipo = Inventario::where("idinventario", "=",$equipo->idinventario )->first();
-        //dd($equipo);
         return view('equipo.show', compact('equipo'));
     }
 
     /**
-     * Show the form for editing the specified resource.
+     * Muesta la forma de edición del equipo.
      *
      * @param  \App\Models\Equipo  $equipo
      * @return \Illuminate\Http\Response
      */
     public function edit(Inventario $equipo)
     {
-        //dd($equipo);
         $numero = null;
         return view('equipo.edit', compact('equipo'), compact('numero') );
     }
 
     /**
-     * Update the specified resource in storage.
+     * Actualiza el equipo en la base de datos.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  \App\Models\Equipo  $equipo
@@ -95,32 +97,58 @@ class EquipoController extends Controller
     public function update(Request $request, $idinventario)
     {
         $Equipo = Inventario::find($idinventario);
+
+        //Revisa si hay una subida de imagen nueva
         if ($request->hasfile('fotoinventario')) {
-            $destination = storage_path('Imagenes\\'.$Equipo->fotoinventario);
-            if (File::exists($destination)) {
-                File::delete($destination);
+            
+            //Si la imagen a cambiar es la defecto, solamente le cambia el nombre
+            if ($Equipo->fotoinventario == 'sinimagen.jpg') {
+                $file = $request->file('fotoinventario');
+                $extention = $file->getClientOriginalExtension();
+                $filename = time().'.'.$extention;
+                $file->move(public_path().'/Imagenes', $filename);
+
+                Inventario::where('idinventario','=', $idinventario)->update([
+                    'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                    'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
+                    'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+                ]);
             }
+            //si no, reemplaza la imagen con la nueva imagen.
+            else{
+                
+                $destination = public_path('Imagenes/'.$Equipo->fotoinventario);
+                if (File::exists($destination)) {
+                    //Lo Elimina
+                    File::delete($destination);
+                }
+                
+                //Realiza la nueva inserccion
+                $file = $request->file('fotoinventario');
+                $extention = $file->getClientOriginalExtension();
+                $filename = time().'.'.$extention;
+                $file->move(public_path().'/Imagenes', $filename);
 
-            $file = $request->file('fotoinventario');
-            $extention = $file->getClientOriginalExtension();
-            $filename = '/'.time().'.'.$extention;
-            $file->move(public_path().'/Imagenes', $filename);
+                Inventario::where('idinventario','=', $idinventario)->update([
+                    'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                    'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
+                    'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+                ]);
+            }
         }
-
-        $datosEquipo = $request->except(['_token','_method']);
-        Inventario::where('idinventario','=', $idinventario)->update([
-            'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
-            'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
-            'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
-        
-        ]);
-
-        //$herramienta = Inventario::findOrFail($idinventario);
+        //Si no, actualiza los otros datos 
+        else{
+            Inventario::where('idinventario','=', $idinventario)->update([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                'tipoinventario'=> $request->tipoinventario, 
+                'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+            ]);
+        }
         return redirect()->route('equipo.index');
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Borra el equipo de la base.
      *
      * @param  \App\Models\Equipo  $equipo
      * @return \Illuminate\Http\Response
@@ -128,11 +156,10 @@ class EquipoController extends Controller
     public function destroy(Inventario $equipo)
     {
         $Equipo = Inventario::find($equipo->idinventario);
-        $destination = storage_path('Imagenes\\'.$Equipo->fotoinventario);
+        $destination = storage_path('Imagenes/'.$Equipo->fotoinventario);
             if (File::exists($destination)) {
                 File::delete($destination);
             }
-        //dd($equipo);
         $equipo->delete();
         return redirect()->route('equipo.index');
     }
