+++ app/Http/Controllers/HerramientaController.php
@@ -10,19 +10,18 @@ use Illuminate\Support\Facades\File;
 class HerramientaController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Muestra el listado de herramientas.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
         $herramientas = Inventario::where('tipoinventario', '=', 'Herramienta')->get();
-        //dd($equipos);
         return view('herramienta.index',compact('herramientas'));
     }
 
     /**
-     * Show the form for creating a new resource.
+     * Muestra la forma para crear una nueva herramienta.
      *
      * @return \Illuminate\Http\Response
      */
@@ -34,64 +33,64 @@ class HerramientaController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Guerda la nueva herramienta en la base de datos.
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
+        //Revisa si hay una imagen insertandose
         if ($request->hasfile('fotoinventario')) {
-
             $file = $request->file('fotoinventario');
             $extention = $file->getClientOriginalExtension();
-            $filename = '/'.time().'.'.$extention;
+            $filename = time().'.'.$extention;
             $file->move(public_path().'/Imagenes', $filename);
-            //dd($filename);
+
+            Inventario::insert([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
+                'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+            ]);
         }
-        $herramienta = $request->except('_token');
-        Inventario::insert([
-            'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
-            'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
-            'estadoinventario' => 'En Buenas Condiciones', 'informacioninventario' => $request->informacioninventario
-        
-        ]);
+        //Si no, se incluye una imagen por defecto para mostrar.
+        else{
+            Inventario::insert([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => 'sinimagen.jpg',
+                'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+            ]);
+        } 
         return redirect()->route('herramienta.index');
-        //dd($equipo);
     }
 
     /**
-     * Display the specified resource.
+     * Muestra una herramienta en específico.
      *
      * @param  \App\Models\Herramienta  $herramienta
      * @return \Illuminate\Http\Response
      */
     public function show(Inventario $herramientum)
     {
-        //$id = strval($idNum);
         $herramienta = Inventario::where("idinventario", "=",$herramientum->idinventario )->first();
-        //dd($herramienta);
-        //$estudiantes = Registro::where("idEs", "=", Estudiante::select("id"));
         return view('herramienta.show', compact('herramienta'));
     }
 
     /**
-     * Show the form for editing the specified resource.
+     * Muestra la forma para actualizar la herramienta.
      *
      * @param  \App\Models\Herramienta  $herramienta
      * @return \Illuminate\Http\Response
      */
     public function edit(Inventario $herramientum)
     {
-        //dd($herramientum);
         $numero = null;
         $herramienta = $herramientum;
-
         return view('herramienta.edit', compact('herramienta'), compact('numero') );
     }
 
     /**
-     * Update the specified resource in storage.
+     * Actualia la herramienta en la base de datos.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  \App\Models\Herramienta  $herramienta
@@ -100,33 +99,55 @@ class HerramientaController extends Controller
     public function update(Request $request, $idinventario)
     {
         $Herramienta = Inventario::find($idinventario);
+        //Revisa si hay una subida de imagen
         if ($request->hasfile('fotoinventario')) {
-            $destination = storage_path('Imagenes\\'.$Herramienta->fotoinventario);
-            if (File::exists($destination)) {
-                File::delete($destination);
+            //Si la imagen a cambiar es la defecto, solamente le cambia el nombre
+            if ($Herramienta->fotoinventario == 'sinimagen.jpg') {
+                $file = $request->file('fotoinventario');
+                $extention = $file->getClientOriginalExtension();
+                $filename = time().'.'.$extention;
+                $file->move(public_path().'/Imagenes', $filename);
+
+                Inventario::where('idinventario','=', $idinventario)->update([
+                    'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                    'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
+                    'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+                ]);
             }
+            //si no, reemplaza la imagen con la nueva imagen
+            else{
+                $destination = public_path('Imagenes/'.$Herramienta->fotoinventario);
+                if (File::exists($destination)) {
+                    //Si hay, lo borra
+                    File::delete($destination);
+                }
 
-            $file = $request->file('fotoinventario');
-            $extention = $file->getClientOriginalExtension();
-            $filename = '/'.time().'.'.$extention;
-            $file->move(public_path().'/Imagenes', $filename);
+                //Inserccion de nueva imagen
+                $file = $request->file('fotoinventario');
+                $extention = $file->getClientOriginalExtension();
+                $filename = time().'.'.$extention;
+                $file->move(public_path().'/Imagenes', $filename);
+            
+                Inventario::where('idinventario','=', $idinventario)->update([
+                    'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                    'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
+                    'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+                ]);
+            }   
         }
-        $datosHerramienta = $request->except(['_token','_method']);
-        //dd($id);
-        Inventario::where('idinventario','=', $idinventario)->update([
-            'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
-            'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
-            'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
-        
-        ]);
-
-        //$herramienta = Inventario::findOrFail($idinventario);
-        //return view('estudiante.edit', compact('estudiante') );
+        //Si no, Sube los otros datos respectivos.
+        else{
+            Inventario::where('idinventario','=', $idinventario)->update([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                'tipoinventario'=> $request->tipoinventario, 
+                'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+            ]);
+        }
         return redirect()->route('herramienta.index');
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Remueve la herramienta de la base de datos.
      *
      * @param  \App\Models\Herramienta  $herramienta
      * @return \Illuminate\Http\Response
@@ -134,11 +155,10 @@ class HerramientaController extends Controller
     public function destroy(Inventario $herramientum)
     {
         $Equipo = Inventario::find($herramientum->idinventario);
-        $destination = storage_path('Imagenes\\'.$Equipo->fotoinventario);
+        $destination = storage_path('Imagenes/'.$Equipo->fotoinventario);
             if (File::exists($destination)) {
                 File::delete($destination);
             }
-        //dd($herramientum);
         $herramientum->delete();
         return redirect()->route('herramienta.index');
     }
