+++ app/Http/Controllers/VehiculoController.php
@@ -10,19 +10,18 @@ use Illuminate\Support\Facades\File;
 class VehiculoController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Muestra un el listado de vehículos.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
         $vehiculos = Inventario::join('vehiculos', 'idinventario', '=', 'vehiculos.idinventariofk')->get();
-        //dd($vehiculos);
         return view('vehiculo.index',compact('vehiculos'));
     }
 
     /**
-     * Show the form for creating a new resource.
+     * Muestra la forma para crear un nuevo vehículo.
      *
      * @return \Illuminate\Http\Response
      */
@@ -34,38 +33,48 @@ class VehiculoController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Guarda un nuevo vehículo en la base de datos.
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
+        //Revisamos si hay una imagen insertandose
         if ($request->hasfile('fotoinventario')) {
-
             $file = $request->file('fotoinventario');
             $extention = $file->getClientOriginalExtension();
-            $filename = '/'.time().'.'.$extention;
+            $filename = time().'.'.$extention;
             $file->move(public_path().'/Imagenes', $filename);
-            //dd($filename);
+    
+            $id = Inventario::insertGetId([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
+                'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+            ]);     
+            Vehiculo::insert([
+                'idinventariofk' => $id, 'patentevehiculo' => $request->patentevehiculo,
+                'tipovehiculo' => $request->tipovehiculo
+            ]); 
         }
-        $id = Inventario::insertGetId([
-            'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
-            'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => $filename,
-            'estadoinventario' => 'En Buenas Condiciones', 'informacioninventario' => $request->informacioninventario]
-            
-        );    
-        //dd($id);    
-        Vehiculo::insert([
-            'idinventariofk' => $id, 'patentevehiculo' => $request->patentevehiculo,
-            'tipovehiculo' => $request->tipovehiculo
-        ]);
-        return redirect()->route('vehiculo.index');
+        //Si no, se incluye una imagen por defecto para mostrar.
+        else {
+            $id = Inventario::insertGetId([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario, 
+                'tipoinventario'=> $request->tipoinventario, 'fotoinventario' => 'sinimagen.jpg',
+                'estadoinventario' => $request->estadoinventario, 'informacioninventario' => $request->informacioninventario
+            ]);     
+            Vehiculo::insert([
+                'idinventariofk' => $id, 'patentevehiculo' => $request->patentevehiculo,
+                'tipovehiculo' => $request->tipovehiculo
+            ]);
+        }
         
+        return redirect()->route('vehiculo.index');
     }
 
     /**
-     * Display the specified resource.
+     * Muestra un vehículo en específico.
      *
      * @param  \App\Models\Vehiculo  $vehiculo
      * @return \Illuminate\Http\Response
@@ -73,14 +82,12 @@ class VehiculoController extends Controller
     public function show(Inventario $vehiculo)
     {
         $vehiculoEleg = Inventario::join('vehiculos', 'idinventario', '=', 'vehiculos.idinventariofk')->where('idinventario', '=', $vehiculo->idinventario)->first();
-        //$herramienta = $herramientum;
         $vehiculo = $vehiculoEleg;
-        //dd($material);
         return view('vehiculo.show', compact('vehiculo'));
     }
 
     /**
-     * Show the form for editing the specified resource.
+     * Muestra la forma para editar un vehículo existente.
      *
      * @param  \App\Models\Vehiculo  $vehiculo
      * @return \Illuminate\Http\Response
@@ -88,15 +95,13 @@ class VehiculoController extends Controller
     public function edit(Inventario $vehiculo)
     {
         $vehiculoEleg = Inventario::join('vehiculos', 'idinventario', '=', 'vehiculos.idinventariofk')->where('idinventario', '=', $vehiculo->idinventario)->first();
-        //$herramienta = $herramientum;
         $vehiculo = $vehiculoEleg;
         $numero = null;
-        //dd($vehiculo);
         return view('vehiculo.edit', compact('vehiculo'), compact('numero') );
     }
 
     /**
-     * Update the specified resource in storage.
+     * Actualiza el vehículo en la base de datos.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  \App\Models\Vehiculo  $vehiculo
@@ -105,33 +110,76 @@ class VehiculoController extends Controller
     public function update(Request $request, $id)
     {
         $Vehiculo = Inventario::find($id);
+        
+        //Comprueba si hay una imagen nueva
         if ($request->hasfile('fotoinventario')) {
-            $destination = storage_path('Imagenes\\'.$Vehiculo->fotoinventario);
-            if (File::exists($destination)) {
-                File::delete($destination);
+            //Si la imagen a cambiar es la defecto, solamente le cambia el nombre
+            if ($Vehiculo->fotoinventario == 'sinimagen.jpg') {
+
+                //Inserccion de la imagen
+                $file = $request->file('fotoinventario');
+                $extention = $file->getClientOriginalExtension();
+                $filename = time().'.'.$extention;
+                $file->move(public_path().'/Imagenes', $filename);
+
+                Inventario::where('idinventario', $id)
+                ->update([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario,  
+                'fotoinventario' => $filename, 'estadoinventario' => $request->estadoinventario, 
+                'informacioninventario' => $request->informacioninventario
+                ]);
+
+                Vehiculo::where('idinventariofk', $id)
+                ->update([
+                'tipovehiculo' => $request->tipovehiculo, 'patentevehiculo' => $request->patentevehiculo
+                ]);
             }
+            //si no, reemplaza la imagen con la nueva imagen.
+            else {
+                $destination = public_path('Imagenes/'.$Vehiculo->fotoinventario);
+                if (File::exists($destination)) {
+                    //Si existe ruta, lo borra
+                    File::delete($destination);
+                } 
+                //Realiza la nueva inserccion
+                $file = $request->file('fotoinventario');
+                $extention = $file->getClientOriginalExtension();
+                $filename = time().'.'.$extention;
+                $file->move(public_path().'/Imagenes', $filename);
+            
+                Inventario::where('idinventario', $id)
+                ->update([
+                    'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario,  
+                    'fotoinventario' => $filename, 'estadoinventario' => $request->estadoinventario, 
+                    'informacioninventario' => $request->informacioninventario
+                ]);
+                
+                Vehiculo::where('idinventariofk', $id)
+                ->update([
+                    'tipovehiculo' => $request->tipovehiculo, 'patentevehiculo' => $request->patentevehiculo
 
-            $file = $request->file('fotoinventario');
-            $extention = $file->getClientOriginalExtension();
-            $filename = '/'.time().'.'.$extention;
-            $file->move(public_path().'/Imagenes', $filename);
+                ]);
+            }            
         }
-        Inventario::where('idinventario', $id)
-        ->update([
-            'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario,  
-            'fotoinventario' => $filename, 'estadoinventario' => $request->estadoinventario, 
-            'informacioninventario' => $request->informacioninventario
-        ]);
-        Vehiculo::where('idinventariofk', $id)
-        ->update([
-            'tipovehiculo' => $request->tipovehiculo, 'patentevehiculo' => $request->patentevehiculo
+        //Si no, actualiza el resto de los datos
+        else{
+            Inventario::where('idinventario', $id)
+            ->update([
+                'codinventario' => $request->codinventario, 'nombreinventario' => $request->nombreinventario,  
+                'estadoinventario' => $request->estadoinventario, 
+                'informacioninventario' => $request->informacioninventario
+            ]);
+            Vehiculo::where('idinventariofk', $id)
+            ->update([
+                'tipovehiculo' => $request->tipovehiculo, 'patentevehiculo' => $request->patentevehiculo
 
-        ]);
+            ]);
+        }
         return redirect()->route('vehiculo.index');
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Borra un vehículo de la lista.
      *
      * @param  \App\Models\Vehiculo  $vehiculo
      * @return \Illuminate\Http\Response
@@ -139,11 +187,10 @@ class VehiculoController extends Controller
     public function destroy(Inventario $vehiculo)
     {
         $Vehiculo = Inventario::find($vehiculo->idinventario);
-        $destination = storage_path('Imagenes\\'.$Vehiculo->fotoinventario);
+        $destination = storage_path('Imagenes/'.$Vehiculo->fotoinventario);
             if (File::exists($destination)) {
                 File::delete($destination);
             }
-        //dd($vehiculo);
         $vehiculo->delete();
         return redirect()->route('vehiculo.index');
     }
